Move mock email to helper util
